chore(board list): moved scenarios from 'MyBoards' to 'Admin'
