+++ next-tavla/pages/edit/boards/[id].tsx
@@ -2,7 +2,7 @@ import classes from 'styles/pages/admin.module.css'
 import { Contrast } from '@entur/layout'
 import { ToastProvider } from '@entur/alert'
 import { Header } from 'components/Header'
-import { Admin } from 'src/MyBoards/scenarios/Admin'
+import { Admin } from 'Admin/scenarios/Boards'
 import { TSettings } from 'types/settings'
 import { checkFeatureFlags } from 'utils/featureFlags'
 import { getBoards } from 'utils/firebase'
